Fixed meas_meta_sir compiler warnings.
Fixed a couple of string formatting errors.  Set strings intended to
hold filenames to stdio.h FILENAME_MAX.  Read a few articles comparing
strcpy to strncpy.  I used to think strcpy was unsafe but now I'm
thinking that strcpy is fine for copying a constant string.  For the
ezdump change, I think it's OK for now, we will eliminate it
eventually. The new netcdf writer will (hopefully) be safer.


Former-commit-id: eef1d8f9b2f2c1d9611954d950ba819569039fa0