Final Resolution fixes
Corrected the size of the grd_a_image in meas_meta_make.  Default for
our project is for it to be 25 km, it was being set incorrectly.
Also modified Makefiles so that on virtual machines, the NSIDC test
directory is on /projects/PMESDR/vagrant/NSIDCtest because there isn't
enough disk space to run the full daily regression test on /home/vagrant.


Former-commit-id: 4660a9ea1a16e773097b3e9660d7deee0db22470