Deleted locally declared strings
cetb_get_swath_producer_id_from_outpath was passing on dev machine but
failing on scm.  I suspect, (and confirmed as well as I could), that
this was due to stack memory having junk on scm that isn't there on dev.
Changed the logic so that locally declared strings are no longer needed.