Created a single meas_meta_setup
Setup has been modified to expect GSX files as input - regardless of CSU
or RSS origin.  (Next mod will be to take ANY GSX file, e.g. AMSRE or
SSMI/S) There is now a single Makefile for setup and both csu_ and rss_
targets are present in this file