Changed sir to write TB instead of a_image
Also edited the compare_cetb code to read either bgi_image or a_image
from the first file and to read the new, scaled TB variable from the
second file.  And to ignore the new GRD files that are being produced.
When we do produce a full new set of regression files, we should remove
the stuff that ignores the GRD files and we should start looking for TB
in any file, regardless of argument position.