Modified Make to note input data source
Modified meas_meta_make to read the cetb_platform_id from the command
line and use that number for the setup files.  Also implemented David's
scheme of F=ssmi, A=amsre, I=ssmi/s and R=smmr for the setup files and
coded the channel id to go from 1-9 and a-wherever for more than 9
channels (as on AMSRE).  Added amsre targets to Makefiles