Fixing PMESDR-200, dimension values wrong on janus
On janus icc compiler, "2.D" in denominator compiles as zero.
Compiles/runs as expected on gcc.  Specifying a constant type is a
FORTRAN habit that is unnecessary in C.  Fixed the problem in cetb_file.
Was causing NaNs in cols/rows dimension variables produced on janus.
Also added a new check to cetb_utilities to compare first element of
cols dimension variable.