Fixed the T grid input file problem
T grids were processing all input files in the list, irrespective of the
date of the scans in the file.  This resulted in multiple swaths from
different days processed onto the T grid.  A check was added to make
sure that for ascending/descending grid processing, only data from the
DOY are used.