End_header is written always
If the list of input files in the meta file is empty, setup will no rite
the End_header line which is a trigger for sir and bgi to stop reading
info and start reading data out of the setup file.  Previously, a
missing End_header line resulted in an infinite loop in these 2 applications.