Change max_diff_pixels for VM
With the addition of some double to float casts, the results of the
regression have changed slightly.  The regressions on Janus still pass
with the limits we set up on that computer.  The regressions on the VM
now require max_diff_pixels to be set to 100 rather than 0 - this seemed
more straightforward and honest than rebuilding the regression files.