TB_std_dev fill value was set incorrectly
TB_std_dev was set to MAX UINT, but, the variable was still a float so
it didn't get tagged as fill when written to the NETCDF file.  Changed
this so that it is now set correctly.  Also set a flag to note if there
is an OOR TB value as this will then require the OOR (missing) value to
be set for TB_std_dev