Janus compile flags to limit warnings
Added compile flags for icc compiler to ignore the unused variable
warnings that are resulting from variables included in cetb.h - they are
not needed in each program in the system.
Also added a use for azang in setup - until we decide how to use it
based on early adopter feedback.