Write box size to setup file
The size of the search box that is used to find measurements that lie
within the threshold for each grid point needs to be written out to the
metadata for the output grid files.  First stage in the process is to
write the box size in km out to the setup file - thereafter it can be
read into sir and written out to the final netCDF file.