Changed nc template to remove unneeded TBD's
There were still 3 TBD's in the global metadata that will not be
modified when the output netcdf files are written, viz. for source, gsx
version and command line history.

We decided that all of the relevant parameters that trace the history of
the output data creation are either part of the TB variable
metadata (e.g. number of iterations, response threshold, LTOD cutoffs
etc) and we have replaced the source and gsx version global meta data
items with number of input files and the list of input files