Fixed several orbit direction problems.
Forcing a setup run to be NS or T but not both.  Fixed/simplified logic
to handle LTOD vs ASCDES logic. Eliminated bug that was setting ascend
flag to D if sequential latitudes were single-precision equal.  We are
now throwing away the first scan of the very first file and keeping
track of last known scantime and latitude.  Making sure that comparison
of scanlines are within a minute, otherwise we don't have enough
information to decide ASCDES.