Small tweaks to regression process.
A few changes that are required for running from scratch as a new user.
git lfs install is needed to configure git to understand lfs repos.
Adjusted the regression data directory for when it doesn't exist to begin with.
Deleted a bunch of old hardcoded copies of regression scripts, in favor of
run_regression.sh.